usb: refactor to hidapi
LibUSB is not ideal in all circumstances as it requires blocking on
each device. HidAPI seems to handle this internally (less work for
the library :) so it seems worth porting to. It also supports things
like serial numbers and such better, although it is a little quirky
in xplatform open() and such.

Daemon read working in this commit, but write() and other features
need some work before this is done.