ctlra: wip on cleanup on exit()
Current state: it works but segfaults on exit. Cause is some mismatch
in the linked-list and the device / application freeing stuff.

Re hotplug; status is that the hotplug function gets called, but the
actual concept of "connect()" vs "accept_connection()" must be reversed.

AKA: we should ask the application does it want to accept VID/PID device,
and if yes, it returns the event_func() and userdata*, which is then set
on the device. This puts the burden of tracking controllers that are the
application is interested in on the application, and allows Ctlra to probe
the application once for each supported controller that is plugged in.

This requires some refactoring of connect(), as well as the general API
to remain consistent in how to use it.