maschine: usb pads investigating
This commit hacks a linux/hidraw backend into the
maschine-mikro driver, replacing libUSB. the result
is significantly better pad message handling (in the
end-user, less-crosstalk way).

Conclusion: libusb (and the current Ctlra usage) is *not*
suitable for pad messages, for an unknown reason.

A guess is that 1) USB Capability Report changes frequency,
which the kernel handles for hidraw, but our driver does not?