devs: remove flush() lights from connect
connect() implementation may not have any USB transactions,
as this is a race condition with setting up the error handling
of the USB backend. This could be avoided by demanding each
connect() function takes a ctlra_t* arg, and sets it before
calling - but connect probably shouldn't do I/O of any kind.

The cleaner fix is to have an init_routine() function if the dev wants
to do fancy lights / show-off-y things.