ctlra: add get_name() function to info_t struct
This allows an application to get the number of controls, and
the various control names from the info struct. As an end result,
the info is exposed in the accept() function not only after accept.

From the application this makes things much easier, as when accepting
a controller we can immidiatly set-up the UI and retrieve the names
of each controller using the info provided function.