ctlra: update info get name() api
This allows the get_name() API call to be made from Accept,
which makes it easier for applications to retrieve names
of controls and cache them prior to events actually arriving.

Simple example updated to use the new API. Drivers need to be
ported to set the function pointer in the info struct.