dev_test: add sample to test devices
Now in a semi-working state - the UI shows up, and displays
the correct one (from accepted dev). Dev events -> UI still
an issue, need to design it.

Keep in mind that jog wheels are *NOT* feedback items, but for
device testing it is useful to be able to have those "feedback"
avialable to update the UI. Its a mismatch between I and O.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>