towards feedback: lot of changes - details inside
This commit is a POC of item feedback using a feedback_id
inside the item. It provides the input item with an identifyer
for output (eg: LED) number to use.

The application must make use of this info - the driver should not
perform any light changes itself.

Also prototyped here is some changes to the device_test app to
make use of the info.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>