ctlra: initial screen feedback item implemented
This commit adds initial infrastructure for the Ctlra library
to represent a screen as a feedback item with a physical layout.

Previous ctlra_dev_screen_get_data() function is how to actually
get the data pointer for the screen drawing, however this may be
totally overhauled in future commits, in order to better support
virtual devices and screens.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>