jam: fix grid event sending handling
This commit fixes the grid input handling, which previously would
send an event for every pressed button (when holding down items, there
can be many pressed at the same time).

With this commit, we check that the value is different to the previous
verion of the grid-button state, which allows identifying when the state
of the grid-square changes, and only emitting a single event when it's
state changes.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>