jam: initial feedback ids added device layout info
This commit adds the feedback ids of the surrounding buttons
of the JAM device.

The internal buttons (1,2,3,4 etc) and the named buttons
(A,B,C,D etc) are on different USB ids, and not yet supported,
neither is the grid of 8x8 buttons. Touch-strip LEDs are similarly
not yet supported.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>