3dconnexion: SpaceMouse Pro support added
This commit adds support to Ctlra for the SpaceMouse Pro.
The wireless version of the device was used for testing,
although it is expected to be similar to the wired version.

All input buttons are mapped to Ctlra events, and the 6-dof
knob controller is mapped to 12 sliders as follows:
- Move Left
- Move Right
- Move Forward
- Move Backward
- Move Up
- Move Down
- Rotate Anti-Clockwise
- Rotate Clockwise
- Rotate Left
- Rotate Right
- Rotate Backward
- Rotate Forward.

Combining two axis back into a single -1..1 range value is trivial
in the application, if the application wants to use that range.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>