usb: rework handles and interfaces, fixing close(), #47
This commit renames the usb_interface, to the more accurate
usb_handle (it stores the libusb handles...) and re-adds
a usb_interface. The interface is an integer that is used
to close the handle later.

With this patch, the closing of devices doesn't spit out
any warnings - assuming no transactions are in-flight,
that work remains to do!

Relates to #47

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>