usb: fix mem leak in interrupt read on error case, #47
This leak was found by valgrind, and later debugged properly
using AddressSanitizer, which pointed directly to the origin
of the alloc that wasn't being freed.

The number of objects leaked was exactly the same as the number
of error read ops, hence it was easy to identify and fix the root
cause of the mem leaks.

Signed-off-by: Harry van Haaren <harry.van.haaren@gmail.com>