ctlra: add digits feedback API
This commit adds a digits feedback API, allowing users
to display float numbers on "7 segment" digit displays.
The x1mk2 interface was used to test this feature, and
it seems to work OK. The driver itself has to convert
the number into digits, and then display them as best
it can.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>