devices/midi: provide feedback through light_set
This commit allows feedback to be sent to any MIDI device,
by re-using the light-set API. The API has two methods of
setting lights, one "generally working" one, which is going
to work for *most* MIDI controllers, and an advanced usage,
which allows all 3 of the MIDI bytes to be provided by the
caller of the API - this should enable any use-case.

The simple usage is as follows, where "light_id" is the MIDI note number
to send a NOTE_ON message to, with the 3rd byte set to "value":

  ctlra_dev_light_set(dev, light_id, value);

The advanced usage is as follows, where "light_id" is set to UINT32_MAX
(aka, -1 as it is unsigned). With this value set, the generic MIDI backend
interprets the "value" as a 3 byte MIDI message, as follows:

  uint8_t msg[3] = {
    0x0, /* status byte here */
    0x0, /* 2nd byte */
    0x0, /* 3rd byte */
  };

  ctlra_dev_light_set(dev, -1, value);

With these two usages, it is expected that almost any device can be
controlled (I have not experienced controllers that use 2 byte messages
to control LEDs. If they exist, please contact me, and we can find a
solution using the unused 8 bits at the top of "value").

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>