ctlra: add pressure to button event
This commit adds pressure to a button event, allowing communication
of more than a basic on/off. This is useful for mapping MIDI backend,
but also for a range of more versatile controllers based on other
protocols.

A new flag is added to the event_button_t, has_pressure. When set,
the event has a pressure value set, and is valid to read. If has_pressure
is not set, the value is garbage and should not be used.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>