avtka: add screen data capabilities
This commit uses the latest AVTKA source, and enables creation
of "screens" on the AVTKA UI. This allows virtualization of the
device screens, providing feedback to developers what *would*
show up on the screen, if they had access to the hardware.

See here for details of the AVTKA changes:
https://github.com/openAVproductions/openAV-Avtka/pull/5

This showcases how an application currently the screen draw APIs.
It should probably be refactored to take into account different
pixel-layout issues, some helper functions provided to ease
development and debugging.

Also working towards abstracting out all screen data so the application
has the knowledge it needs to draw to any screen (in theory..)

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>