ctlra: fix return if invalid result in virtual dev
This commit fixes a possible segfault if a controller
is virtualized, and the device returns NULL instead of
a valid device_t from its connect() function.

Previously a NULL pointer + offset would be passed to the
application accept_dev_func(), which would often result
in a segfault. Returning early before the app dev accept
call handles this.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>