maschine3: large rework of pads and cleanup
This commit adds initial support for pads - although nowhere
near production quality yet. Median filter to be used to detect
release events, which will allow the device's "addition" of a
pad to the "pressed list" be the trigger for note-on events.
This should provide much more accurate note onsets, and provide
a nice slider for when a release occurs (pressure wise).

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>