ctlra: rework screen register API
This commit reworks the register function to register a single
callback for each device, and the callback has a parameter stating
the screen idx that this call is for. This simplifies handling in
the Ctlra library, and the application can easily switch on the
screen idx if they wish to handle each screen individually.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>