ctlra/event: rework for grid/tcc compatibility
Note this is not a good solution yet - the exact binary
sizes and offsets must be specified to ensure other compilers
adhere to the exact same binary struture.

This will become more important if a SHM based approach is
taken to share controllers - ultimately the event + ABI must
remain consistent unless breakage between Ctlra daemon and
app is to occur.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>