usb: fix segfault on immidiate bulk xfer on connect
This commit fixes a possible segfault if a device with
bulk endpoints attempts to do bulk transfers on connect(),
when the device is not yet ready to accept them. This results
in a LIBUSB_ERROR_BUSY return from the transfer, and would
previously cause the device to be banished. However the banish
call accessed a NULL pointer for the dev->ctlra_context, causing
the segfault.

Fixed by handling ERROR_BUSY and not banishing the device,
and adding a TODO to pass the ctlra_context to the connect()
call so the ctlra_context pointer is always valid.