ctlra: rework API for dev setting callbacks
This initial rework cleans up the "accept" function in
each application. Instead of writing to a bunch of
function-pointer pointers (yes that's correct), we now
provide a handle that can be used to register any and all
callbacks for the device in question.

As a bonus, in any callback these same functions can be used
to update/change/reset the callback functions themselves.

Note this commit is a prototype API only, and needs implementation
and testing with a complex sample (eg: loopa).

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>