ctlra: add set_callback_userdata() function
This function allows the application to set a void*
on the device, and this will be passed to each of the
callback functions for: event, feedback, screen, remove.

Simple applications might not make use of it, but TCC
enabled complex applications (e.g. Mixxx) make use of
a different TCC instance per device, and that can be
passed around using this void *.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>