ctlra-cairo: cleanup and implement optimized conversion code
This commit implements cairo-565 to device-565 encoding. A simple
byte swap is required, which is costly in scalar code due to the
large number of loads and stores, combined with data-dependencies
due to the shifts.

The optimized routines use SIMD registers to achieve the same effect,
but operating on multiple pixels at once, and only using one load/store
for the set of pixels.

Overall, the previously ~12% of CPU time on the scalar implmentation
has been reduced to ~0.68 on the SSE version, and ~0.32% for AVX2.

Currently the optimized routines are only used if compiled in - no
runtime detection of CPU features.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>