mappa: initial commit of v1 input mapping API
This commit adds a sample application that implements an
initial draft of the mappa API. In particular the abstraction
between the application and the mappa software is being kept
as simple as possible - even still it will probably change
a few times - so do not start using this API just yet!

This commit closes #80 initial mapping design.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>