mappa: implemented event lut, and event handling
This commit is a significant step towards POC of using Ctlra
look-up-table for event mapping. The software target events
are stored in a linear array sorted by type, so the lookup of
the target is just an array access. The function pointer is
called if one is set.

This almost completes #81.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>