mappa: refactor error handling, no more crashy
This commit strenghtens the error checking on events as compared
to the events that are advertised by the device's metadata. If
the event id is beyond the advertised range, it is ignored and
a print issued to notify - this is due to invalid event/metadata
pairing in the device code.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>