mappa: implement lookup APIs for ID
This commit adds an API to lookup a target_id from name.
This should complete the basics required to load a config
file, and map any controller to targets based on the file.

Basic layer switching is only supported now - in future
it must be expanded to more flexible methods.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>