mappa: wip conversion to ctlra name based config
This commit is a ~40% conversion towards using control names
as the config file items, instead of slider.0 and slider.1 etc.

This will improve the human-readable aspect of the mappa config
hugely, and also simplifies generation of bindings in a mapping
editor (no string<->int conversion between target and slider id).

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>