mappa: large refactor of config file loading
This commit refactors a large amount of the config file loading,
scraping the item names from the Ctlra device metadata. The config
file is now human readable "key = value" pairs, with "key" being
a Ctlra device driver control name, and "value" being the target
string of the software to map to.

"Key" is defined by the code implemented in the Ctlra driver. The
names of the controls *MUST* refect the hardware device. If no
logical name exists, a self-descriptive must be used to identify
that singular control. Changing button names is *NOT* permitted,
as doing so approximates an ABI break for users' config files.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>