mappa: mapping_count error check value added
The "mapping_count" item in a layer allows the human to indicate
the number of mappings that are created for this layer. This allows
mappa to check that exactly that number of mappings are made. If
fewer are made, warning/error prints are emitted to inform the
user that something has gone wrong.

WIP on active-on-load, which should enable activating layers once
the file is loaded.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>