mappa: initial overlay list implementation
This commit adds the concept of the active_lut_list, which contains
the currently active layers in the order that they should be handled.
The entries in this lut_list must be itereated and flattended into the
dev->active_lut in order for them to be used for event input lookup.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>