mappa: layer active switching using names working
This is a WIP commit but does prove the concept of the layer
active switching working with the sample .ini file.

This shows that the mapping can be achieved using the overlay
techinque, and that the button mapping to the layer-switch target
allows the user to program all aspects of multilayered mappings
using just the .ini config file, and some knowledge of the
controller and software in question.

Note that with virtual devices, a software developer can easily
assist the user with any issues that arise - actually the development
of this commit (and most of the previous mappa related ones) was
done entirely using virtual devices, away from the hardware. With
the hardware present, the cross-over was seamless to have working
input mappings with those devices.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>