mappa: implement value scaling and set_range() api
This commit implements value scaling inside Mappa. The end
result is that an application can set a range for each target,
which Mappa then automatically range-scales and offsets any
float value mapped to it.

The end result is "transparent" range-scaling for users, where
a full movement of the hardware translates to full movement of
the software parameter (aka; basic expected experience).

However the value of having the target range exposed here is that
mappa can in future support "ranges" for each control, which can
be user defined, and the range numbers are not 0-1 based, but actual
real units (eg, remap lowpass filter min value from 0Hz to 110Hz, so
the sound is never 100% cut out).

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>