mappa: rework how config files are used by app
This commit refactors the previous "load_bindings()" function,
replacing it with the more generic add_config_file(). This allows
the application to indicate config files that are present, and
gives mappa freedom to scan/load them as appropriate on hotplug
of devices.

All in all, this reduces complexity in the application (eg: no
need to implement a callback to get the correct mappa config for
a specific devices). It also allows future mappa to timestamp
loads of files, and auto-reload files etc.

Note that this commit breaks API/ABI of load_bindings()

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>