mappa: Remove add_config_file() API
This commit removes the add_config_file API, as the
software should not directly control the location of
mappa files - they should be centralized so it is easy
to find and copy-paste bits from others.

Might re-add this in future to allow applications to
add thier own specific directory - but for now it doesn't
seem smart to *require* that apps use this.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>