mappa: feedback mappings to lights working
This approach requires a "sources" array for each LUT/device
to contain the maximum size for any light_id of the device.
A source_t represents the source of that particular value to
display, and callbacks are triggered when we need to update.

This approach requires that Ctlra drivers expose the max
number of endpoints that can be valid-ly called on the device.
Aka, a device with 1000 leds will return 1000. This info is
not yet available in Ctlra for each device, but will be required
to properly support this feature.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>