mappa: rework screens to be part of LUTs
This change allows the screen UIs to be set by each LUT,
giving the user the flexibility to have a button activate
a layer, and hence also change the display.

The combination of one LUT having the input/lights/screen
is a good one, to allow easier UX of developing/testing mappings.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>