s4mk1: initial commit, see issue #94
This is an initial input-only implementation for the S4 Mk1,
which uses a USB Bulk based protocol. It transfers data using
the USB Bulk endpoints, with multiple message "blocks" inside
a single bulk transfer.

The buttons and sliders should be functioning correctly with
this commit, however it has not been fully tested yet.

Future work includes adding the jog-wheel support, validating
the inputs (buttons/sliders), and enabling LED support.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>