s4mk1: add warning and next steps re usb access
This commit adds a warning to the code that indicates the
current status of the S4Mk1 work. The sliders and other
parameters are working, however detaching from the kernel
driver is almost certainly not what end-users will want,
given the audio-card on the S4 Mk1 enables turntables and
monitoring while playing etc.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>