ctlra: add screen redraw request API for event driven redraws
This allows the application to determine when a screen redraw is
performed. In reality, it is just noted that a redraw request has
been made, and defers the actual redrawing to the next ctlra_idle_iter()
call.

For lowest latency, applications should call idle_iter() directly
after the redraw request.

Device polling and redraw is not implemented internally in request
redraw as the application may still be processing events from another
thread (eg. through a ringbuffer). As there may be more events
outstanding after the redraw-causing event, it makes sense to let the
application decide when to call idle_iter().

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>